refactor/rename some parts of the codes
