+++ pcdet/models/backbones_3d/focal_sparse_conv/focal_sparse_conv.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 import spconv.pytorch as spconv
 from pcdet.ops.roiaware_pool3d.roiaware_pool3d_utils import points_in_boxes_gpu
-from pcdet.models.backbones_3d.focal_sparse_conv.utils import split_voxels, check_repeat, FocalLoss
+from pcdet.models.backbones_3d.focal_sparse_conv.focal_sparse_utils import split_voxels, check_repeat, FocalLoss
 from pcdet.utils import common_utils
 
 
@@ -221,5 +221,4 @@ class FocalSparseConv(spconv.SparseModule):
         out = out.replace_feature(self.bn1(out.features))
         out = out.replace_feature(self.relu(out.features))
 
-        batch_dict['loss_box_of_pts'] += loss_box_of_pts
-        return out, batch_dict
+        return out, batch_dict, loss_box_of_pts
